feat: Add header along with menu and menu background pattern and self logo SVGs
The code changes include adding two new SVG files, `menu-bg-pattern.svg` and `self-logo.svg`, to the `public` directory. These SVGs are used to enhance the visual appearance of the website.

In addition, the `layout.tsx` file has been modified to import and use a new `Header` component. This component will be displayed on the website's layout.

Lastly, a new CSS file, `header.css`, has been added to the `app/components` directory. This file contains styles for the header element, including a gradient mask and a background pattern.